% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{find_git_root_file}
\alias{find_git_root_file}
\title{File paths relative to the root of a directory hierarchy}
\usage{
find_git_root_file(..., path = ".")
}
\arguments{
\item{...}{\verb{[character]}\cr
Further path components passed to \code{\link[=file.path]{file.path()}}.
All arguments must be the same length or length one.}

\item{path}{\verb{[character(1)]}\cr
The start directory.}
}
\description{
A shortcut of \code{rprojroot::find_root_file} for git repositories.
}
\seealso{
\code{\link[rprojroot:find_root_file]{rprojroot::find_root_file}}
}
