% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_sized_clustering.R
\name{fixed_size_clustering_nearest_neighbor}
\alias{fixed_size_clustering_nearest_neighbor}
\alias{fixed_size_clustering_kmeans}
\alias{fixed_size_clustering_hclust}
\title{Create clusters of fixed size}
\usage{
fixed_size_clustering_nearest_neighbor(d, size, method = c("max", "min"))

fixed_size_clustering_kmeans(x, size, method = c("range", "min", "max"))

fixed_size_clustering_hclust(d, size, method = "ward.D2")
}
\arguments{
\item{d}{\code{dist}, dissimilarity structure as produced by \code{dist}.}

\item{size}{\code{integer}, fixed target size of the clusters.}

\item{method}{\code{character}, start with the closest (min)/farthest (max) point in average.}

\item{x}{\code{numeric}/\code{matrix}, points to be clustered.}
}
\value{
\code{integer}, vector of the same length as \code{attr(d, "Size")}
with cluster identification numbers.
}
\description{
Create clusters of fixes size using the nearest neighbor approach.
}
\details{
This algorithm runs \verb{[stats::kmeans]} once to determine the center position
and reassigns the points to the nearest center with respect to the cluster
size. The \code{kmeans} could run multiple times to update every center after a
cluster was filled but we used the current approach for simplicity.

\code{methods}: \code{min} (\code{max}), start with the closest (farthest) points. \code{range}
start with the points where the difference in farthest minus nearest distance
is largest.

\code{fixed_size_clustering_hclust}:
This algorithm runs \verb{[stats::hclust()]} once to determine the clustering
and cuts the tree at different heigths to ensure the fixed cluster sizes.

\code{methods}: \code{ward.D2} clustering method, see \verb{[stats::hclust()]} for details.
}
\examples{
x <- c(1:3, 1e4, 11:13, 1.1e4, 111:113)
d <- dist(x)
fixed_size_clustering_nearest_neighbor(d, size = 3)
fixed_size_clustering_kmeans(d, size = 3)
fixed_size_clustering_hclust(d, size = 3)
}
\references{
Algorithm taken from Jean Monlong
\url{https://github.com/jmonlong/Hippocamplus/blob/master/content/post/2018-06-09-ClusterEqualSize.Rmd}

kmeans Algorithm inspired from from Jean Monlong
\url{https://github.com/jmonlong/Hippocamplus/blob/master/content/post/2018-06-09-ClusterEqualSize.Rmd}
and \url{https://elki-project.github.io/tutorial/same-size_k_means}
}
