% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_sized_clustering.R
\name{fixed_sized_clustering}
\alias{fixed_sized_clustering}
\title{Create clusters of fixed size}
\usage{
fixed_sized_clustering(d, size)
}
\arguments{
\item{d}{\code{dist}, dissimilarity structure as produced by \code{dist}.}

\item{size}{\code{integer}, fixed target size of the clusters.}
}
\value{
\code{integer}, vector of the same length as \code{attr(d, "Size")}
with cluster identification numbers.
}
\description{
Create clusters of fixes size using the nearest neighbor approach.
}
\examples{
x <- c(1:3, 1e4, 11:13, 1.1e4, 111:113)
d <- dist(x)
fixed_sized_clustering(d, size = 3)
}
