% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-clustering.R
\name{total_sum_of_squares}
\alias{total_sum_of_squares}
\alias{within_cluster_sum_of_squares}
\title{Calculate sum of squares}
\usage{
total_sum_of_squares(x)

within_cluster_sum_of_squares(x, cl)
}
\arguments{
\item{x}{\code{matrix}}

\item{cl}{\code{integer}, vector with cluster ids, length has to be equal
\verb{nrow(x))}.}
}
\value{
\code{double} sum of squares
}
\description{
Calculate sum of squares
}
\examples{
total_sum_of_squares(as.matrix(1:3))
m <- matrix(1:4, nrow = 4)
within_cluster_sum_of_squares(m, c(1, 1, 1, 2))
}
