% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_sized_clustering.R
\name{fixed_sized_clustering}
\alias{fixed_sized_clustering}
\title{Create clusters of fixed size}
\usage{
fixed_sized_clustering(d, size, method = c("max", "min"))
}
\arguments{
\item{d}{\code{dist}, dissimilarity structure as produced by \code{dist}.}

\item{size}{\code{integer}, fixed target size of the clusters.}

\item{method}{\code{character}, start with the closest (min)/farthest (max) point in average.}
}
\value{
\code{integer}, vector of the same length as \code{attr(d, "Size")}
with cluster identification numbers.
}
\description{
Create clusters of fixes size using the nearest neighbor approach.
}
\examples{
x <- c(1:3, 1e4, 11:13, 1.1e4, 111:113)
d <- dist(x)
fixed_sized_clustering(d, size = 3)
}
\references{
Algorithm taken from Jean Monlong
\url{https://github.com/jmonlong/Hippocamplus/blob/master/content/post/2018-06-09-ClusterEqualSize.Rmd}
}
